==> 4-vaCleaning.Rmd <==
---
title: "3-vaCleaning"
author: "Will Sheppard"
output: html_document
date: "2023-02-07"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)

# import functions
source('functions_visionProcessing_and_threshEstimation.R')
```


```{r read data}
vaOnline <- read.csv("../data/visionTestsOnline.csv")%>%
  filter(display == 'va')
vaLab <- read.csv("../data/visionTestsLab.csv")%>%
  filter(display == 'va')
vaChart <- read.csv("../data/visualTestData.csv")%>%
  select(id, vaChart)
```

```{r shorten va}
vaShortOnline <- shortenVA(vaOnline)
vaShortLab <- shortenVA(vaLab)
```

we do not remove scores with long RTs here, many conditions only have 1 response so we lose thsee data


```{r create summary data frames}
vaSummOnline <- summShortVA(vaShortOnline)
vaSummLab <- summShortVA(vaShortLab)

```

save out summary data for psychometric functions
```{r save summ data}

write.csv(vaSummLab, "../data/vaSummLab.csv")
write.csv(vaSummOnline, "../data/vaSummOnline.csv")

```

lab particiapnt 56 missing
online = 6 missing
```{r calc thresholds}
vaThreshOnline <- threshVAround(vaSummOnline, 'Online') 
vaThreshLab <- threshVAround(vaSummLab, 'Lab')
```

```{r merge data frames}

vaThresh <- merge(vaThreshOnline, vaThreshLab)%>%
  select(-ends_with('RT'))
vaThresh <- merge(vaThresh, vaChart)
```

```{r save data}
write.csv(vaThresh, "../data/vaThreshClean.csv")
```
