==> 13-exp1_discussion.Rmd <==
---
title: "exp1_discussion"
author: "Will Sheppard"
date: "2023-12-04"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

libraries
```{r libraries}

require(tidyverse)
```


```{r read data}

vaImport <- read.csv("../onlineVAtestComparison.csv")

```


va comparison
```{r va}

va <- vaImport %>%
  mutate(absBias = round(abs(BIAS), digits = 2),
         loaWidth = round((ULOA - LLOA)/2, digits = 2))%>%
  filter(Reference.Test != "LeedsVis US")

va$ours <- as.factor(va$ours)

#not normally distributed
shapiro.test(va$absBias)

wilcox.test(va$absBias ~ va$ours)

ggplot(va, aes(y = absBias, group = ours, color = ours))+
  geom_boxplot()
ggplot(va, aes(y = loaWidth, group = ours, color = ours))+
  geom_boxplot()

cols <- c("1" = "red", "0" = "black")
shapes <- c("1" = 3, "0" = 16)

vaOthers <- va %>%
  filter(Reference != "Sheppard et al (2024)")

mWidth <- mean(vaOthers$loaWidth, na.rm = T)
mBias <- mean(vaOthers$absBias, na.rm = T)

vaComp <- ggplot(va, aes(x = absBias, y = loaWidth))+
  geom_point(aes(shape = ours), size = 3)+
  scale_shape_manual(values = shapes, labels = c("Other studies", "Current study"))+
  xlab("Absolute bias (logMAR)")+
  ylab("LOA width (logMAR)") +
  geom_hline(yintercept = mWidth, linetype = "dashed")+
  geom_vline(xintercept = mBias, linetype = "dashed")+
  theme_bw()+
  theme(legend.title = element_blank(),
        legend.position = c(0.9, 0.11))

show(vaComp)
```

```{r va save}

dpi = 300
h = 4
w = 7

ggsave("../figures/SUBMISSION/figure3.tiff",
       plot = vaComp,
       dpi = dpi,
       width = w,
       height = h)
```
